cmd/snap-confine: move device cgroup support to libsnap-confine-private
